+++ packages/core/src/provider/plugins/sass.ts
@@ -22,6 +22,7 @@ export function pluginSass(): RsbuildPlugin {
         const { excludes, options } = getSassLoaderOptions(
           config.tools.sass,
           // source-maps required for loaders preceding resolve-url-loader
+          // otherwise the resolve-url-loader will throw an error
           true,
         );